requirements.txt
altgraph==0.17.4
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
docxcompose==1.4.0
docxtpl==0.16.7
et-xmlfile==1.1.0
idna==3.6
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
MouseInfo==0.1.3
openpyxl==3.1.2
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-docx==1.1.0
pytweening==1.0.7
pywin32-ctypes==0.2.2
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
